Begin StreamReader c++ implementation - create base .hpp/.cpp files
StreamReader object contains a thread that runs on member function. The member
function updates an int passed into the StreamReader constructor. While loop
in member function updates the int variable and checks for a termination flag
(i.e. StreamReader object destructor is called). StreamReader runs only with an
int and is not attached to the dependency tree.